requirements.txt
streamlit>=1.40.0
python-pptx>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
